Calculations for example 4 are done
Created a function to do calculations based on example 4.
Aspect receives a score as the mean from only the most polarized
words present in opinion/related columns. This means at maximum
two words contribute to the score. If a related or opinion word is
missing the the aspect receives automatically the most polarized score
from the remaining column.